FUTURE_ML_03/requirements.txt
streamlit
python-dotenv
openai
google-cloud-dialogflow
python-telegram-bot==13.15
textblob
pandas